% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advarsel.R
\name{get_advarsel}
\alias{get_advarsel}
\title{Get and parse the alarmeringsapp rss feed}
\usage{
get_advarsel(rss_feed = "https://alarmeringsapp.like.st/rss")
}
\arguments{
\item{rss_feed}{is the url to the advarsels rss feed}
}
\value{
a data frame of the ten latest
}
\description{
Get and parse the alarmeringsapp rss feed
}
\examples{
library(alarmeringsapp)
alarms <- get_advarsel()
}
